implements RRT-Connect in BasicRRTStrategy.h and introduces the parameter growGoals in the XML file which determines whether or not to grow the goal as well as the start nodes.  If using a directed graph, do not grow the goal.  I also modified the equal operator in Cfg.cpp and the RRTExpand method in MPUtils.h in order to compute when two trees meet.  Approved by Shawna (correctness) and Andy (standards).
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@4001 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Cfg/Cfg.cpp
@@ -99,11 +99,11 @@ Cfg::operator==(const Cfg& _cfg) const {
     return false;
   for(size_t i = 0; i < m_dof; ++i) {
     if(m_dofTypes[i] == POS || m_dofTypes[i] == JOINT) {
-      if(fabs(m_v[i]-_cfg[i]) > numeric_limits<double>::epsilon())
+      if(fabs(m_v[i]-_cfg[i]) > numeric_limits<float>::epsilon())
         return false;
     }
     else {
-      if(fabs(DirectedAngularDistance(m_v[i], _cfg[i])) > numeric_limits<double>::epsilon()) 
+      if(fabs(DirectedAngularDistance(m_v[i], _cfg[i])) > numeric_limits<float>::epsilon()) 
         return false;
     }
   }
